Change type in SalesManager int to long

+++ src/Main.java
@@ -1,9 +1,9 @@
 public class Main {
 
     public static void main(String[] args) {
-        int[] arr = {30, 90, 150};
+        long[] arr = {30, 90, 150};
         SalesManager salesManager = new SalesManager(arr);
-        int max = salesManager.max();
+        long max = salesManager.max();
         System.out.println(max);
     }
 }
